Simple 2D Animation
This second project was to create a 2D animation. I made this as a conceptual animation of a hypothetical "alligator" cell that could be created as a way to target and destroy cells in the bloodstream by binding and performing the "alligator death roll" on them to rip up their cell membranes.