3D Animation of ATP Synthase
This animation is a 3D animation meant to simulate the behavior of an ATP synthase. One bug that currently exists is that it appears that the random drift of the protons can lead to them diffusing across the membrane to a certain point. This is purely caused by z-axis drift and only occurs because I couldn't animate a 3D membrane without blocking the view. (One solution would simply be to limit the z-axis drift)